# LSP Optimization: Use rayon for `parse_ast_to_typed_tokens` function