… sway-lsp/src/traverse/typed_tree.rs