@JoshuaBatty 2024-01-17
> How come collecting typed tokens is taking `127.7ms`? Or does that figure include parsing and type checking already?

This is how long it's taking to traverse and collect the tokens from the typed AST nodes. I'm going to look at using rayon par_iters for this section like we are doing in the `traverse::parsed_tree` module. Should hopefully take us down closer to `12ms` or so.